requirements.txt
Django==5.0.6
djangorestframework==3.15.1
youtube-search-python==1.6.6
python-dotenv==1.0.1
llama-index-readers-google
llama-index==0.10.37
chromadb
psycopg2
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface